Fixed UserData

+++ src/screen.h
@@ -15,7 +15,7 @@ int GetEncoderClsId(const WCHAR *format, CLSID *clsId);
 void CaptureScreen(const std::string &path,
 				   const std::string &name,
 				   bool loopForever = false,
-				   int delay = 0,
+				   float delay = 0.0f,
 				   int loopAmount = 0);
 }
